Split engine registry into smaller modules
