+++ app/engines/bridge_utils.py
@@ -1,0 +1,50 @@
+"""Shared utilities for the voice bridge."""
+
+from __future__ import annotations
+from typing import Any
+from app.engines.errors import EngineRequestError
+
+INTENDED_UPSTREAM_CALLERS = (
+    "app.domain.voices.preview",
+    "app.orchestration.scheduler.orchestrator",
+    "app.orchestration.tasks",
+)
+INTENDED_DOWNSTREAM_DEPENDENCIES = (
+    "app.engines.registry.load_engine_registry",
+    "app.engines.voice.base.BaseVoiceEngine",
+    "app.engines.tts_client.TtsClient",
+)
+FORBIDDEN_DIRECT_IMPORTS = (
+    "app.api.routers",
+    "app.db",
+    "app.jobs",
+)
+
+
+def extract_engine_id(request: dict[str, Any]) -> str:
+    """Extract and validate engine_id from a request dictionary."""
+    engine_id = str(request.get("engine_id") or "").strip()
+    if not engine_id:
+        raise EngineRequestError("Voice requests must include engine_id.")
+    return engine_id
+
+
+def extract_synthesis_settings(request: dict[str, Any]) -> dict[str, Any]:
+    """Extract per-request synthesis settings for the TTS Server path."""
+    settings: dict[str, Any] = {}
+    if "voice_profile_id" in request:
+        settings["voice_profile_id"] = request["voice_profile_id"]
+    if "voice_asset_id" in request:
+        settings["voice_asset_id"] = request["voice_asset_id"]
+    if "speed" in request:
+        settings["speed"] = request["speed"]
+    if "safe_mode" in request:
+        settings["safe_mode"] = request["safe_mode"]
+    if "output_format" in request:
+        settings["output_format"] = request["output_format"]
+    return settings
+
+
+def infer_audio_format(output_path: str) -> str:
+    """Infer audio format from output path extension."""
+    return "mp3" if output_path.lower().endswith(".mp3") else "wav"
